+++ stroom-core-client/src/main/java/stroom/data/client/presenter/DataPresenter.java
@@ -36,6 +36,7 @@ import stroom.pipeline.shared.AbstractFetchDataResult;
 import stroom.pipeline.shared.FetchDataAction;
 import stroom.pipeline.shared.FetchDataResult;
 import stroom.pipeline.shared.FetchMarkerResult;
+import stroom.pipeline.shared.SourceLocation;
 import stroom.pipeline.shared.StepLocation;
 import stroom.security.client.api.ClientSecurityContext;
 import stroom.security.shared.PermissionNames;
@@ -86,13 +87,14 @@ public class DataPresenter extends MyPresenterWidget<DataPresenter.DataView> imp
     private SharedList<Marker> markers;
     private int startLineNo;
     private List<Highlight> highlights;
-    private Long highlightStreamId;
+    private Long highlightMetaId;
     private Long highlightStreamNo;
-    private String highlightChildStreamType;
+    private String highlightChildDataType;
     private boolean playButtonVisible;
     private ClassificationUiHandlers classificationUiHandlers;
     private BeginSteppingHandler beginSteppingHandler;
     private boolean steppingSource;
+    private boolean formatOnLoad;
 
     @Inject
     public DataPresenter(final EventBus eventBus, final DataView view, final TextPresenter textPresenter,
@@ -178,7 +180,7 @@ public class DataPresenter extends MyPresenterWidget<DataPresenter.DataView> imp
 
     @Override
     public void beginStepping() {
-        if (beginSteppingHandler != null) {
+        if (beginSteppingHandler != null && currentMetaId != null) {
             beginSteppingHandler.beginStepping(currentMetaId, currentChildDataType);
         }
     }
@@ -194,6 +196,64 @@ public class DataPresenter extends MyPresenterWidget<DataPresenter.DataView> imp
 
         this.currentChildDataType = childDataType;
         update(true);
+    }
+
+    public void fetchData(final SourceLocation sourceLocation) {
+        this.highlights = new ArrayList<>();
+        this.highlights.add(sourceLocation.getHighlight());
+
+        this.highlightMetaId = sourceLocation.getStreamId();
+        this.highlightStreamNo = sourceLocation.getStreamNo();
+        this.highlightChildDataType = sourceLocation.getChildStreamType();
+
+        // Set the source type that will be used when the page source is shown.
+        // Make sure the right page is shown when the source is displayed.
+        final long oldStreamNo = currentDataRange.getOffset() + 1;
+        long newStreamNo = oldStreamNo;
+        final long oldPageOffset = currentPageRange.getOffset();
+        final long pageLength = currentPageRange.getLength();
+        int lineNo = 0;
+
+        // If we have a source highlight then use it.
+        if (highlights != null && highlights.size() > 0) {
+            final Highlight highlight = highlights.get(0);
+            newStreamNo = highlightStreamNo;
+            lineNo = highlight.getFrom().getLineNo();
+        }
+
+        final long page = lineNo / pageLength;
+        long newPageOffset = oldPageOffset % pageLength;
+        final long tmp = page * pageLength;
+        if (tmp + newPageOffset < lineNo) {
+            // We can show this page.
+            newPageOffset = tmp + newPageOffset;
+        } else {
+            // We need to show the page before.
+            newPageOffset = tmp - pageLength + newPageOffset;
+        }
+
+        // Update the stream source.
+        if (!EqualsUtil.isEquals(currentMetaId, highlightMetaId)
+                || !EqualsUtil.isEquals(currentChildDataType, highlightChildDataType) || oldStreamNo != newStreamNo
+                || oldPageOffset != newPageOffset) {
+            currentDataRange = new OffsetRange<>(newStreamNo - 1, 1L);
+            currentPageRange = new OffsetRange<>(newPageOffset, pageLength);
+
+            this.currentMetaId = highlightMetaId;
+            this.currentChildDataType = highlightChildDataType;
+            dataTypeOffsetRangeMap.clear();
+            markerListPresenter.resetExpandedSeverities();
+//            currentStreamRange = new OffsetRange<>(sourceLocation.getStreamNo() - 1, 1L);
+//            currentPageRange = new OffsetRange<>((long) sourceLocation.getHighlight().getFrom().getLineNo() - 1, 1L);
+//            streamTypeOffsetRangeMap.clear();
+//            markerListPresenter.resetExpandedSeverities();
+
+//            fetchData(false, highlightStreamId, highlightChildStreamType);
+            update(false);
+        } else {
+            refreshHighlights(lastResult);
+            refreshMarkers(lastResult);
+        }
     }
 
     public void fetchData(final boolean fireEvents, final Long metaId, final String childDataType) {
@@ -405,7 +465,7 @@ public class DataPresenter extends MyPresenterWidget<DataPresenter.DataView> imp
         pageRows.updateRowData(pageOffset, pageLength);
         pageRows.updateRowCount(pageCount, pageCountExact);
 
-        textPresenter.setText(data, true);
+        textPresenter.setText(data, formatOnLoad);
         textPresenter.setFirstLineNumber(startLineNo);
         textPresenter.setControlsVisible(playButtonVisible);
 
@@ -414,18 +474,18 @@ public class DataPresenter extends MyPresenterWidget<DataPresenter.DataView> imp
     }
 
     private void refreshHighlights(final AbstractFetchDataResult result) {
-        int streamOffset = 0;
+        int streamNo = 0;
 
         if (result != null) {
-            streamOffset = result.getStreamRange().getOffset().intValue();
+            streamNo = result.getStreamRange().getOffset().intValue() + 1;
         }
 
         // Make sure we have a highlight section to add and that the stream id
         // matches that of the current page, and that the stream number matches
         // the stream number of the current page.
-        if (highlights != null && currentMetaId != null && currentMetaId.equals(highlightStreamId)
-                && streamOffset == highlightStreamNo
-                && EqualsUtil.isEquals(currentChildDataType, highlightChildStreamType)) {
+        if (highlights != null && currentMetaId != null && currentMetaId.equals(highlightMetaId)
+                && streamNo == highlightStreamNo
+                && EqualsUtil.isEquals(currentChildDataType, highlightChildDataType)) {
             // Set the content to be displayed in the source view with a
             // highlight.
             textPresenter.setHighlights(highlights);
@@ -451,13 +511,13 @@ public class DataPresenter extends MyPresenterWidget<DataPresenter.DataView> imp
     public void showStepSource(final Integer taskOffset, final StepLocation stepLocation,
                                final String childStreamType, final List<Highlight> highlights) {
         this.highlights = highlights;
-        this.highlightStreamId = stepLocation.getStreamId();
+        this.highlightMetaId = stepLocation.getStreamId();
         this.highlightStreamNo = stepLocation.getStreamNo() - 1;
-        this.highlightChildStreamType = childStreamType;
+        this.highlightChildDataType = childStreamType;
 
         // Set the source type that will be used when the page source is shown.
         // Make sure the right page is shown when the source is displayed.
-        final long oldStreamNo = currentDataRange.getOffset();
+        final long oldStreamNo = currentDataRange.getOffset() + 1;
         long newStreamNo = oldStreamNo;
         final long oldPageOffset = currentPageRange.getOffset();
         final long pageLength = currentPageRange.getLength();
@@ -467,7 +527,7 @@ public class DataPresenter extends MyPresenterWidget<DataPresenter.DataView> imp
         if (highlights != null && highlights.size() > 0) {
             final Highlight highlight = highlights.get(0);
             newStreamNo = highlightStreamNo;
-            lineNo = highlight.getLineFrom();
+            lineNo = highlight.getFrom().getLineNo();
         }
 
         final long page = lineNo / pageLength;
@@ -482,13 +542,13 @@ public class DataPresenter extends MyPresenterWidget<DataPresenter.DataView> imp
         }
 
         // Update the stream source.
-        if (!EqualsUtil.isEquals(currentMetaId, highlightStreamId)
-                || !EqualsUtil.isEquals(currentChildDataType, highlightChildStreamType) || oldStreamNo != newStreamNo
+        if (!EqualsUtil.isEquals(currentMetaId, highlightMetaId)
+                || !EqualsUtil.isEquals(currentChildDataType, highlightChildDataType) || oldStreamNo != newStreamNo
                 || oldPageOffset != newPageOffset) {
-            currentDataRange = new OffsetRange<>(newStreamNo, 1L);
+            currentDataRange = new OffsetRange<>(newStreamNo - 1, 1L);
             currentPageRange = new OffsetRange<>(newPageOffset, pageLength);
 
-            fetchData(false, highlightStreamId, highlightChildStreamType);
+            fetchData(false, highlightMetaId, highlightChildDataType);
         } else {
             refreshHighlights(lastResult);
             refreshMarkers(lastResult);
@@ -506,6 +566,10 @@ public class DataPresenter extends MyPresenterWidget<DataPresenter.DataView> imp
     public void setSteppingSource(final boolean steppingSource) {
         this.steppingSource = steppingSource;
         errorMarkerMode = !steppingSource;
+    }
+
+    public void setFormatOnLoad(final boolean formatOnLoad) {
+        this.formatOnLoad = formatOnLoad;
     }
 
     public interface DataView extends View {