+++ stroom-core-client/src/main/java/stroom/hyperlink/client/HyperlinkType.java
@@ -4,5 +4,7 @@ public enum HyperlinkType {
     TAB,
     DIALOG,
     DASHBOARD,
+    STEPPING,
+    DATA,
     BROWSER
 }