+++ stroom-pipeline/src/main/java/stroom/pipeline/filter/SplitFilter.java
@@ -17,15 +17,20 @@
 package stroom.pipeline.filter;
 
 import org.xml.sax.Attributes;
+import org.xml.sax.Locator;
 import org.xml.sax.SAXException;
 import stroom.pipeline.factory.ConfigurableElement;
 import stroom.pipeline.factory.PipelineProperty;
 import stroom.pipeline.shared.ElementIcons;
 import stroom.pipeline.shared.data.PipelineElementType;
 import stroom.pipeline.shared.data.PipelineElementType.Category;
+import stroom.pipeline.state.LocationHolder;
+import stroom.pipeline.state.MetaHolder;
 import stroom.pipeline.xml.event.Event;
 import stroom.pipeline.xml.event.simple.StartElement;
 import stroom.pipeline.xml.event.simple.StartPrefixMapping;
+
+import javax.inject.Inject;
 
 /**
  * Splits a single XML instance into separate XML instances that contain
@@ -108,6 +113,24 @@ public class SplitFilter extends AbstractXMLFilter {
      */
     private boolean hasFiredEvents;
 
+    private final MetaHolder metaHolder;
+    private final LocationHolder locationHolder;
+    private Locator locator;
+    private boolean storeLocations = true;
+    private SourceLocationFactory sourceLocationFactory;
+
+    public SplitFilter() {
+        this.metaHolder = null;
+        this.locationHolder = null;
+    }
+
+    @Inject
+    SplitFilter(final MetaHolder metaHolder,
+                final LocationHolder locationHolder) {
+        this.metaHolder = metaHolder;
+        this.locationHolder = locationHolder;
+    }
+
     @Override
     public void endProcessing() {
         super.endProcessing();
@@ -126,8 +149,17 @@ public class SplitFilter extends AbstractXMLFilter {
         afterRoot = -1;
         depth = 0;
         count = 0;
+        if (sourceLocationFactory != null) {
+            sourceLocationFactory.reset();
+        }
 
         super.startStream();
+    }
+
+    @Override
+    public void setDocumentLocator(final Locator locator) {
+        super.setDocumentLocator(locator);
+        this.locator = locator;
     }
 
     /**
@@ -282,6 +314,10 @@ public class SplitFilter extends AbstractXMLFilter {
         }
 
         if (depth == splitDepth) {
+            if (storeLocations && locationHolder != null && metaHolder != null) {
+                getSourceLocationFactory().storeLocation();
+            }
+
             buffer = true;
 
             if (count == splitCount || afterRoot < 0) {
@@ -289,6 +325,7 @@ public class SplitFilter extends AbstractXMLFilter {
                 count = 0;
                 // Fire all end events.
                 fireEndEvents(length, 0);
+
                 super.endDocument();
             } else {
                 // If the count has not yet reached the split count then we only
@@ -298,6 +335,17 @@ public class SplitFilter extends AbstractXMLFilter {
         }
 
         depth--;
+    }
+
+    private SourceLocationFactory getSourceLocationFactory() {
+        if (sourceLocationFactory == null) {
+            if (locator.getClass().getSimpleName().contains("DS")) {
+                sourceLocationFactory = new DSSourceLocationFactory(metaHolder, locationHolder, locator, splitCount);
+            } else {
+                sourceLocationFactory = new XmlSourceLocationFactory(metaHolder, locationHolder, locator, splitCount);
+            }
+        }
+        return sourceLocationFactory;
     }
 
     /**
@@ -394,5 +442,13 @@ public class SplitFilter extends AbstractXMLFilter {
             displayPriority = 2)
     public void setSplitCount(final int splitCount) {
         this.splitCount = splitCount;
+    }
+
+    @PipelineProperty(
+            description = "Should this split filter store processing locations.",
+            defaultValue = "true",
+            displayPriority = 3)
+    public void setStoreLocations(final boolean storeLocations) {
+        this.storeLocations = storeLocations;
     }
 }