+++ stroom-pipeline/src/main/java/stroom/pipeline/stepping/SteppingTaskHandler.java
@@ -480,6 +480,7 @@ class SteppingTaskHandler extends AbstractTaskHandler<SteppingTask, SteppingResu
 
         // Set the source meta.
         metaHolder.setMeta(meta);
+        metaHolder.setChildStreamType(streamTypeName);
 
         try {
             final StreamLocationFactory streamLocationFactory = new StreamLocationFactory();
@@ -487,20 +488,20 @@ class SteppingTaskHandler extends AbstractTaskHandler<SteppingTask, SteppingResu
 
             // Determine which stream number to start with.
             final long count = source.count();
-            long index = 1;
+            long streamNo = 1;
             if (currentLocation != null) {
                 // If stream no has been set beyond the last stream no then
                 // start at the end.
                 if (currentLocation.getStreamNo() > count) {
                     // Start at the last stream number.
-                    index = count;
+                    streamNo = count;
                     // Update the current processing location.
-                    currentLocation = new StepLocation(meta.getId(), index, currentLocation.getRecordNo());
+                    currentLocation = new StepLocation(meta.getId(), streamNo, currentLocation.getRecordNo());
                 } else {
                     // Else start at the current location.
-                    index = currentLocation.getStreamNo();
+                    streamNo = currentLocation.getStreamNo();
                     // Update the current processing location.
-                    currentLocation = new StepLocation(meta.getId(), index, currentLocation.getRecordNo());
+                    currentLocation = new StepLocation(meta.getId(), streamNo, currentLocation.getRecordNo());
                 }
             }
 
@@ -511,17 +512,17 @@ class SteppingTaskHandler extends AbstractTaskHandler<SteppingTask, SteppingResu
             // sequentially. Loop over the stream boundaries and process
             // each sequentially until we find a record.
             boolean done = controller.isFound();
-            while (!done && index > 0 && index <= count && !Thread.currentThread().isInterrupted()) {
+            while (!done && streamNo > 0 && streamNo <= count && !Thread.currentThread().isInterrupted()) {
                 // Set the stream number.
-                metaHolder.setStreamNo(index - 1);
-                streamLocationFactory.setStreamNo(index);
+                metaHolder.setStreamNo(streamNo);
+                streamLocationFactory.setStreamNo(streamNo);
 
                 // Process the boundary making sure to use the right
                 // encoding.
                 controller.clearAllFilters(null);
 
                 // Get the stream.
-                try (final InputStreamProvider inputStreamProvider = source.get(index - 1)) {
+                try (final InputStreamProvider inputStreamProvider = source.get(streamNo - 1)) {
                     metaHolder.setInputStreamProvider(inputStreamProvider);
                     final SizeAwareInputStream inputStream = inputStreamProvider.get(streamTypeName);
 
@@ -552,17 +553,17 @@ class SteppingTaskHandler extends AbstractTaskHandler<SteppingTask, SteppingResu
                         // If we are stepping forward increment the stream
                         // number, otherwise decrement the stream number.
                         if (StepType.FIRST.equals(stepType)) {
-                            index++;
-                            currentLocation = new StepLocation(meta.getId(), index, 0);
+                            streamNo++;
+                            currentLocation = new StepLocation(meta.getId(), streamNo, 0);
                         } else if (StepType.BACKWARD.equals(stepType)) {
-                            index--;
-                            currentLocation = new StepLocation(meta.getId(), index, Long.MAX_VALUE);
+                            streamNo--;
+                            currentLocation = new StepLocation(meta.getId(), streamNo, Long.MAX_VALUE);
                         } else if (StepType.FORWARD.equals(stepType)) {
-                            index++;
-                            currentLocation = new StepLocation(meta.getId(), index, 0);
+                            streamNo++;
+                            currentLocation = new StepLocation(meta.getId(), streamNo, 0);
                         } else if (StepType.LAST.equals(stepType)) {
-                            index--;
-                            currentLocation = new StepLocation(meta.getId(), index, Long.MAX_VALUE);
+                            streamNo--;
+                            currentLocation = new StepLocation(meta.getId(), streamNo, Long.MAX_VALUE);
                         }
                     }
                 } catch (final IOException | RuntimeException e) {