+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+* Issue **#1244** : Updated Dropwizard to version 1.3.14 to fix session based memory leak.
+
+* Issue **#1246** : Removed elastic search document type, menu items and filter.
+
+* Issue **#1247** : Added XSLT functions (`location`, `streamId`, `streamNo`, `recordNo`, `lineFrom`, `colFrom`, `lineTo`, `colTo`) to determine the current source location so it can be embedded in a cooked event. Events containing raw source location info can be made into links in dashboard tables or the text pane so that a user can see raw source data or jump directly to stepping that raw record. 
+
 * Add data retention feature and index optimisation to Solr indexes.
 
 * Initial support for Solr indexing and search.
@@ -13,8 +19,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * Issue **#1244** : Updated Dropwizard to version 1.3.14 to fix session based memory leak.
 
 * Issue **#1246** : Removed elastic search document type, menu items and filter.
-
-* Backport of change from 6.1 that protects index shard searches finishing prematurely. 
 
 * Issue **#1214** : Fixed issue where the max results setting in dashboard tables was not always being obeyed. Also fixed some dashboard table result page size issues.